Fix cancellation of pending tasks
Formerly, all uncancelled tasks have been treated as if they are done.
This resulted in exceptions when trying to fetch the result (or
exception) from the task for logging purposes.

Now, all tasks are cancelled if they are not already cancelled or not
done.

Tests have been added for cancelling active and queued tasks as well as
coroutines since this wasn't adequately tested previously.