Refactor dropping a client, resolves #81
Formerly, it was possible that a client, who is currently in the
process of being dropped, could still forward messages to other clients.

In order to fix this, we have to store the running task loops on the
client, so the receive (and forward) loop is being cancelled once the
client is being dropped.