Fix accidental task loop cancellation
When the task loop awaited a task that has been cancelled, it re-raised
the `asyncio.CancelledError` and thereby accidentally cancelled itself.
With this fix, we ensure that the task loop cannot be cancelled since
we simply discard cancellation errors. Due to the way the task loop is
structured, it will eventually exit.